Refactor(config): Migrate LLM configuration and factory to new domain structure
- Deleted legacy LLM configuration module and associated factory class to streamline the codebase.
- Created new LLM domain structure under app/llm/ with config.py and factory.py.
- Updated main settings to remove dependency on the old LLMConfig.
- Adjusted imports in pipeline modules to reference LLMFactory from the new location.
- Implemented @lru_cache for singleton pattern in the new config.
- Enhanced maintainability and reduced complexity in LLM provider management.
- Verified functionality of all components post-refactor.
