+++ app/config/settings.py
@@ -6,16 +6,11 @@ Main settings for the application using Pydantic Settings.
 
 from functools import lru_cache
 
-from pydantic import Field
 from pydantic_settings import BaseSettings, SettingsConfigDict
-
-from app.config.llm_config import LLMConfig
 
 
 class Settings(BaseSettings):
     """Main settings for the application."""
-
-    llm: LLMConfig = Field(default_factory=LLMConfig)
 
     model_config = SettingsConfigDict(env_file=".env", extra="ignore")
