+++ app/pipeline/event/lookup/extractor.py
@@ -11,8 +11,8 @@ from typing import Any
 from app.core.exceptions import ErrorMessages, LLMServiceError
 from app.core.node import Node
 from app.core.schema.task import TaskContext
+from app.llm.factory import LLMFactory
 from app.pipeline.schema.lookup import LookupContext, LookupResponse
-from app.services.llm_factory import LLMFactory
 from app.services.logger_factory import logger
 from app.services.prompt_loader import PromptManager
 from app.utils.datetime_utils import get_datetime_reference
